Implementation of -p option: show packet index in -m and -M modes
Sometimes it's usefull to know from which bgp packet output line was generated.
My particular case was about joining bgp attributes with bgp output, but it
can be usefull in other cases.

If -p is given in -m or -M modes, bgpdump will put a packet index into
second column, which is an integer, incremented every input block (which is packet
in case of update files and prefix for table_dumps).